pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentshield"
version = "0.1.0"
description = "AgentShield Spending Agent Firewall"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "sqlmodel>=0.0.24",
  "sqlalchemy>=2.0.43",
  "psycopg[binary]>=3.2.9",
  "redis>=6.4.0",
  "httpx>=0.28.1",
  "pydantic-settings>=2.10.1",
  "alembic>=1.16.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "ruff>=0.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
include = ["app*"]

